Change speed based on class (#85)

+++ core/src/com/beverly/hills/money/gang/screens/data/PlayerConnectionContextData.java
@@ -28,7 +28,7 @@ public class PlayerConnectionContextData {
   @NonNull
   private final Integer movesUpdateFreqMls;
   @NonNull
-  private final Integer speed;
+  private final Float speed;
   @NonNull
   private final Map<Weapon, WeaponStats> weaponStats;
   @NonNull
